refactor(checklist): extract ChatChecklist to own file
